Factor out get_current_listing function so tasks.py and serializers.py can share it
