+++ api/tasks.py
@@ -1,15 +1,7 @@
 from tracker import celery_app
 from api.models import User, Tracker
+from core_listing_scraper import get_current_listings
 from email_services.email_utils import send_email_for_new_or_updated_listings
-
-import json
-import subprocess
-
-import redis
-r = redis.StrictRedis(host='redis', port=6379, db=0)
-
-PYTHON = 'python'
-MAIN_SCRAPY_SCRIPT = 'core_listing_scraper/initiate_spider.py'
 
 @celery_app.task(name='api.update_trackers')
 def update_trackers():
@@ -37,19 +29,6 @@ def update_trackers():
 
             tracker.save()
 
-def get_current_listings(results_page_url):
-    subprocess.call([PYTHON, MAIN_SCRAPY_SCRIPT, results_page_url])
-
-    data = {}
-    listings = json.loads(r.get(results_page_url), encoding='utf-8')
-    for listing in listings:
-            data[listing['craig_id']] = {
-                'title': listing['title'],
-                'price': listing['price'],
-                'absolute_url': listing['absolute_url'],
-                'last_modified_at': listing['last_modified_at']
-            }
-    return data
 
 def get_new_or_updated_listings(outdated_listings, current_listings):
